Move grid list up in tree
You already have a list component, which does not do anything but an item component, which shows both the list and the items. I moved the list up to the grid component and the item now displays just one item. I also added Proptypes so that the props are correct. You hsould use these as well or typescirpt.